Add localization switch within Preferences
Added a combo box inside the Preferences/General form in order to allow the user to switch the localization. This will translate the application the next time it runs to the language the user selected.

The choice gets saved in a QSettings property. Every time Notepadqq runs, it will translate the tr sections based on the language found in the property, or default to English if no property is found.

Upon the user making a language selection in the Localization combo box, a popup will notify the user of the requirement of a Notepadqq restart for the changes to take effect.

The list of translations in the combo box gets populated by reading the .qm files and extracting the language from them.