feat: Begin/End Selection (closes #393) (#727)
Implemented the Notepad++ Begin/End Selection feature. It allows a user to place the cursor where they want to start a selection and mark it.Then they can use Go-to-line or any other method to get to the end point they want, then mark the end position to have the full range of text be selected.

This avoids the usual issue of losing your selection and initial cursor position when using Go-to-line.